more frontend modularizing
