+++ frontend/src/helpers/scripts.js
@@ -22,15 +22,3 @@ export const updateDownloadLink = (data) => {
 export const updateScroll = () => {
     window.scrollTo(0,document.body.scrollHeight);
 }
-
-export const copyToClipboard = () => {
-    const textarea = document.createElement('textarea');
-    textarea.id = 'temp_element';
-    textarea.style.height = 0;
-    document.body.appendChild(textarea);
-    textarea.value = document.getElementById('reportContainer').innerText;
-    const selector = document.querySelector('#temp_element');
-    selector.select();
-    document.execCommand('copy');
-    document.body.removeChild(textarea);
-}
